Fixes on the cuboid selections.

+++ src/main/java/com/mmiillkkaa/randoms/util/Cuboid.java
@@ -1,6 +1,7 @@
 package com.mmiillkkaa.randoms.util;
 
 import org.bukkit.Location;
+import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Block;
 
@@ -11,12 +12,12 @@ import java.util.Random;
  * Assumes x1 < x2, y1 < y2, and z1 < z2
  */
 public class Cuboid {
-    double x1, x2, y1, y2, z1, z2;
+    int x1, x2, y1, y2, z1, z2;
 
     /**
      * A cuboid from the given coordinates.
      */
-    public Cuboid(double x1, double y1, double z1, double x2, double y2, double z2) {
+    public Cuboid(int x1, int y1, int z1, int x2, int y2, int z2) {
         this.x1 = x1;
         this.y1 = y1;
         this.z1 = z1;
@@ -29,7 +30,8 @@ public class Cuboid {
      * A cuboid from the coordinates of the given locations.
      */
     public Cuboid(Location location1, Location location2) {
-        this(location1.getX(), location1.getY(), location1.getZ(), location2.getX(), location2.getY(), location2.getZ());
+        this((int) location1.getX(), (int) location1.getY(), (int) location1.getZ(),
+                (int) location2.getX(), (int) location2.getY(), (int) location2.getZ());
     }
 
     /**
@@ -38,9 +40,9 @@ public class Cuboid {
     public Location getRandomLocation(World world) {
         Random random = new Random();
         Location locationBase = new Location(world, x1, y1, z1);
-        int maxAddToX = Math.abs((int) (x2-x1)) + 1; //Adding one to prevent 0s, which random.nextInt cannot handle
-        int maxAddToY = Math.abs((int) (y2-y1)) + 1;
-        int maxAddToZ = Math.abs((int) (z2-z1)) + 1;
+        int maxAddToX = Math.abs((x2-x1)) + 1; //Adding one to prevent 0s, which random.nextInt cannot handle
+        int maxAddToY = Math.abs((y2-y1)) + 1;
+        int maxAddToZ = Math.abs((z2-z1)) + 1;
         return locationBase.add(random.nextInt(maxAddToX), random.nextInt(maxAddToY), random.nextInt(maxAddToZ));
     }
 
@@ -50,9 +52,9 @@ public class Cuboid {
     public Block[] getBlocks(World world) {
         Block[] blockArray = new Block[getVolume()];
         int i = 0;
-        for(int x = (int) x1; x < x2; x++) {
-            for(int y = (int) y1; y < y2; y++) {
-                for(int z = (int) z1; z < z2; z++) {
+        for(int x = x1; x < x2; x++) {
+            for(int y = y1; y < y2; y++) {
+                for(int z = z1; z < z2; z++) {
                     blockArray[i++] = world.getBlockAt(x, y, z);
                 }
             }
@@ -64,6 +66,6 @@ public class Cuboid {
      * @return The volume of the cuboid.
      */
     public int getVolume() {
-        return (int) (Math.abs((x2 - x1)) * Math.abs((y2 - y1)) * Math.abs((z2 - z1)));
+        return Math.abs((x2 - x1)) * Math.abs((y2 - y1)) * Math.abs((z2 - z1));
     }
 }
